Ev4 Final/requirements.txt
mysql-connector-python
requests
Pillow